requirements.txt
azure-ai-documentintelligence==1.0.2
azure-identity==1.24.0
openai==1.74.0
python-dotenv==1.1.1
streamlit==1.49.0
Pillow==10.4.0